refactor form validation
