+++ app/components/button_visibility_component.html.erb
@@ -1,4 +1,4 @@
-<button class="custom-focus absolute right-4 mt-2 md:mt-4" type="button" data-action="signin#toggleVisibility" >
+<button class="custom-focus absolute right-4 mt-2 md:mt-4" type="button" data-action="validate-form#toggleVisibility" >
   <span class="sr-only">Show password</span>
   <%= image_tag "visibility_on.svg", class: "hidden visibility_on", id: 'visibility_on' %>
   <%= image_tag "visibility_off.svg", class: "visibility_off", id: 'visibility_off' %>
