+++ app/javascript/controllers/signin_controller.js
@@ -2,78 +2,8 @@ import { Controller } from "@hotwired/stimulus"
 
 // Connects to data-controller="signin"
 export default class extends Controller {
-  connect() {
-    this.element.addEventListener("input", this.checkForm);
-  }
 
-  checkForm = () => {
-    if (this.validateEmail() && this.validatePassword()) {
-      error.innerText = '';
-      submitButton.className = "h-[44px] rounded-md text-white w-full bg-black font-bold text-white outline-none cursor-pointer active:ring-2 active:ring-offset-2 active:outline-none custom-focus hover:bg-hover active:bg-hover"
-    }
-  }
-
-  validateEmail() {
-    const emailRegex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/
-    const isValid = emailRegex.test(email.value);
-    this.showValidationForTarget('email', isValid)
-    return isValid
-  }
-
-  validatePassword() {
-    const isValid = password.value.length >= 8
-    this.showValidationForTarget('password', isValid)
-    return isValid
-  }
-
-  findVisibilityOff() {
-    return this.element.querySelectorAll('[class*="visibility_off"]');
-  }
-
-  findVisibilityOn() {
-    return this.element.querySelectorAll('[class*="visibility_on"]');
-  }
-
-  toggleVisibility(event) {
-    if (password.type === "password") {
-      this.findVisibilityOff().forEach(img => img.classList.add('hidden'));
-      this.findVisibilityOn().forEach(img => img.classList.remove('hidden'));
-      password.type = "text";
-      if (typeof confirm_password !== "undefined") {
-        confirm_password.type = "text";
-      }
-    } else {
-      this.findVisibilityOff().forEach(img => img.classList.remove('hidden'));
-      this.findVisibilityOn().forEach(img => img.classList.add('hidden'));
-      password.type = "password";
-      if (typeof confirm_password !== "undefined") {
-        confirm_password.type = "password";
-      }
-    }
-  }
-   
-  showValidationForTarget(target, isValid) {
-    let errorMessage = ''
-
-    if (target === 'email') {
-      errorMessage = "Invalid email address.";
-    }
-    if (target === 'password') {
-      errorMessage = "Password must be at least 8 characters long.";
-    }
-
-    if (isValid) {
-      error.classList.remove("visible")
-      error.classList.add("invisible")
-    } else {
-      error.innerText = errorMessage;
-      error.classList.add("visible");
-      error.classList.remove("invisible")
-      submitButton.className = "h-[44px] rounded-md text-white w-full bg-light-gray font-bold text-white outline-none cursor-not-allowed";
-    }
-  }
-
-  validateForm(event) {
+  submit(event) {
     event.preventDefault();
 
     fetch('https://rs-blackmarket-api.herokuapp.com/api/v1/users/sign_in', {
