+++ app/views/users/login.html.erb
@@ -3,20 +3,20 @@
     <div class="pb-5 pt-[40px] md:p-[47px] md:pb-7">
       <%= render(LogoComponent.new(title: 'Login')) %>
     </div>
-    <%= form_with data: { controller: "signin" } do |form| %>
+    <%= form_with data: { controller: "signin validate-form", action: "submit->signin#submit" } do |form| %>
       <div class="my-4 flex flex-col">
         <%= form.label :email, class: "text-sm md:text-base" %>
-        <%= form.text_field :email, "data-signin-target": "email", class: "border text-black custom-focus h-[34px] w-auto rounded-md border border-dark-violet ps-3 placeholder:text-base placeholder:text-dark-gray hover:border-hover active:outline-none active:ring-2 active:ring-offset-2 md:h-[44px]", placeholder: "Type your email", type: "email" %>
+        <%= form.text_field :email, class: "border text-black custom-focus h-[34px] w-auto rounded-md border border-dark-violet ps-3 placeholder:text-base placeholder:text-dark-gray hover:border-hover active:outline-none active:ring-2 active:ring-offset-2 md:h-[44px]", placeholder: "Type your email", type: "email" %>
       </div>
       <div class="relative my-4">
         <%= form.label :password, class: "text-sm md:text-base" %>
         <div class="flex flex-col relative">
-          <%= form.password_field :password, autocomplete: true, "data-signin-target": "password", class: "border custom-focus h-[34px] w-auto rounded-md border border-dark-violet ps-3 placeholder:text-base placeholder:text-dark-gray hover:border-hover active:outline-none active:ring-2 active:ring-offset-2 md:h-[44px]", placeholder: "Type your password" %>
+          <%= form.password_field :password, autocomplete: true, class: "border custom-focus h-[34px] w-auto rounded-md border border-dark-violet ps-3 placeholder:text-base placeholder:text-dark-gray hover:border-hover active:outline-none active:ring-2 active:ring-offset-2 md:h-[44px]", placeholder: "Type your password" %>
           <%= render(ButtonVisibilityComponent.new()) %>
         </div>
       </div>
       <div class="flex flex-col">
-        <%= form.submit "Log in", data: { action: "signin#validateForm" }, id: "submitButton", class: "h-[44px] rounded-md text-white w-full bg-light-gray font-bold text-white outline-none cursor-not-allowed active:ring-2 active:ring-offset-2 active:outline-none custom-focus" %>
+        <%= form.submit "Log in", id: "submitButton", class: "h-[44px] rounded-md text-white w-full bg-light-gray font-bold text-white outline-none cursor-not-allowed active:ring-2 active:ring-offset-2 active:outline-none custom-focus" %>
       </div>
       <div class="invisible text-red-500 text-center" id="error"></div>
     <% end %>
